Use absolute path in scheduled_task
This is required because the scheduled_task resource insists upon it.
That makes this change a bug-fix. We are exploring whether or not it is
appropriate to relax the checks in the upstream module,
puppetlabs-scheduled_task, for sitautions where a command AND a
working_dir are specified, but until then an absolute path is required
here for the module to work at all.