Expose executable param, fix testing
This allows the puppet_executable for both posix and windows to be set
via the class declaration which, in turn, facilitates this module being
used within another module.

A change is also included that removed an unneeded include line recently
added to the posix and windows files and, instead, puppet_run_scheduler
is included in each class's preconditions.